% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macdown-package.R
\docType{package}
\name{macdown-package}
\alias{macdown}
\alias{macdown-package}
\title{macdown: An updated R Markdown thesis template using the bookdown package}
\description{
This is a modified version of the 'thesisdown' package. It uses the bookdown package and LaTeX graduate thesis templates from McMaster University, this package aims to encourage reproducible research using simple Markdown syntax while embedding all of the 'R' code to produce plots and analyses as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://paezha.github.io/macdown/}
}

}
\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
\keyword{internal}
