% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macdown.R
\docType{package}
\name{macdown}
\alias{macdown}
\alias{macdown-package}
\title{macdown: A package for creating Graduate McMaster theses
using R Markdown}
\description{
This is a modified version of the thesisdown package. It uses the 
    bookdown package and LaTeX graduate thesis templates from McMaster University, 
    this package aims to encourage reproducible research using simple Markdown 
    syntax while embedding all of the `R` code to produce plots and analyses as well.
}
\section{thesis_gitbook}{

Creates an R Markdown thesis template as a webpage
}

\section{thesis_pdf}{

Creates an R Markdown thesis template as a PDF
}

\section{thesis_word}{

Creates an R Markdown thesis template as a Microsoft Word document
}

\section{thesis_epub}{

Creates an R Markdown thesis template as an ebook
}

\author{
\strong{Maintainer}: Antonio Páez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

}
